requirements.txt
Flask
Flask-SQLAlchemy
GeoAlchemy2
psycopg2-binary
ipython
flask-shell-ipython
Flask-Migrate
